refactor: 공통 로직 Piece에 추상 메서드로 분리
