+++ src/main/java/domain/piece/Elephant.java
@@ -2,22 +2,18 @@ package domain.piece;
 
 import domain.Country;
 import domain.Direction;
-import domain.Position;
 import java.util.List;
 
 public class Elephant extends Piece {
+    private static final int ELEPHANT_DIRECTION_SIZE = 3;
+
     public Elephant(Country country) {
         super(new PieceInfo(PieceType.ELEPHANT, country));
     }
 
     @Override
-    public List<Direction> findDirections(Position from, Position to) {
-        List<Integer> distances = from.calculateDistance(to);
-        int x = distances.get(0);
-        int y = distances.get(1);
-
-        List<Direction> directions = Direction.findDirections(x, y);
-        if (directions.size() != 3) {
+    public void validateDirections(List<Direction> directions) {
+        if (directions.size() != ELEPHANT_DIRECTION_SIZE) {
             throw new IllegalArgumentException("[ERROR] 상이 이동할 수 있는 방향은 3개이어야 합니다.");
         }
         if (directions.get(1) != directions.get(2)) {
@@ -26,6 +22,5 @@ public class Elephant extends Piece {
         if (directions.getFirst().isDialog() || !directions.get(1).isDialog()) {
             throw new IllegalArgumentException("[ERROR] 상의 1번째 방향은 직선이고, 2, 3번째 방향은 대각선이어야 합니다.");
         }
-        return directions;
     }
 }
